Move complex tests to separate folder
